move the proccess for visualization
